djang3ByExampleBook/testDj/requirements.txt
django_taggit==1.2.0
markdown==3.1.1
psycopg2-binary==2.8.4